fix loff_t to size_t

+++ aesd-char-driver/aesd-circular-buffer.h
@@ -53,8 +53,9 @@ struct aesd_circular_buffer
 
 extern size_t aesd_circular_buffer_get_length(struct aesd_circular_buffer *buffer);
 
-extern loff_t aesd_circular_buffer_find_seekto_fpos(struct aesd_circular_buffer *buffer, uint32_t write_cmd,
+extern size_t aesd_circular_buffer_find_seekto_fpos(struct aesd_circular_buffer *buffer, uint32_t write_cmd,
                                                     uint32_t write_cmd_offset);
+
 
 extern struct aesd_buffer_entry *aesd_circular_buffer_find_entry_offset_for_fpos(struct aesd_circular_buffer *buffer,
                                                                                  size_t char_offset, size_t *entry_offset_byte_rtn);
